% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_equal_pdf.R
\name{compare_pdf_content}
\alias{compare_pdf_content}
\title{Support PDF snapshotting contents}
\usage{
compare_pdf_content(old, new)
}
\arguments{
\item{old, new}{Paths to old and new snapshot files.}
}
\description{
Support PDF snapshotting contents
}
